Fixes invalid cast exception on Postgres Count
The postgres count operation returns a long. Attempting to cast
directly from type object to int throws an invalid cast exception.
Casting first to a long, then casting to an int gets around this
error without changing the Massive API.

+++ Massive.PostgreSQL.cs
@@ -695,7 +695,7 @@ namespace Massive.PostgreSQL
         }
         public int Count(string tableName, string where = "")
         {
-            return (int)Scalar("SELECT COUNT(*) FROM " + tableName + " " + where);
+            return (int)(long)Scalar("SELECT COUNT(*) FROM " + tableName + " " + where);
         }
 
         /// <summary>
